Fixed a bug When reading a forward slash & CleanUp

+++ source/common.h
@@ -26,7 +26,7 @@ struct myString
 {
 	u8* data;
 	u64 count;
-	u8& operator[]  (int i) {
+	u8& operator[]  (u64 i) {
 		assert(i < count);
 		return data[i];
 	}
